+++ lib/providers/services/navigation_service.dart
@@ -28,10 +28,10 @@ import 'package:seeds/screens/app/wallet/transfer/transfer_form.dart';
 import 'package:seeds/screens/onboarding/join_process.dart';
 import 'package:seeds/screens/onboarding/onboarding.dart';
 import 'package:seeds/v2/screens/explore/explore_screen.dart';
-import 'package:seeds/v2/screens/profile/support/support_screen.dart';
-import 'package:seeds/v2/screens/settings/edit_name/edit_name_screen.dart';
-import 'package:seeds/v2/screens/settings/set_currency/set_currency_screen.dart';
-import 'package:seeds/v2/screens/settings/settings_screen.dart';
+import 'package:seeds/v2/screens/profile_screens/support/support_screen.dart';
+import 'package:seeds/v2/screens/profile_screens/settings/edit_name/edit_name_screen.dart';
+import 'package:seeds/v2/screens/profile_screens/settings/set_currency/set_currency_screen.dart';
+import 'package:seeds/v2/screens/profile_screens/settings/settings_screen.dart';
 import 'package:seeds/widgets/page_not_found.dart';
 
 class Routes {
