+++ lib/v2/datasource/remote/model/profile_model.dart
@@ -25,6 +25,34 @@ class ProfileModel {
     this.timestamp,
   });
 
+  ProfileModel copyWith({
+    String account,
+    String status,
+    String type,
+    String nickname,
+    String image,
+    String story,
+    String roles,
+    String skills,
+    String interests,
+    int reputation,
+    int timestamp,
+  }) {
+    return ProfileModel(
+      account: account ?? this.account,
+      status: status ?? this.status,
+      type: type ?? this.type,
+      nickname: nickname ?? this.nickname,
+      image: image ?? this.image,
+      story: story ?? this.story,
+      roles: roles ?? this.roles,
+      skills: skills ?? this.skills,
+      interests: interests ?? this.interests,
+      reputation: reputation ?? this.reputation,
+      timestamp: timestamp ?? this.timestamp,
+    );
+  }
+
   factory ProfileModel.fromJson(Map<String, dynamic> json) {
     return ProfileModel(
       account: json['account'],
