+++ lib/v2/screens/explore/interactor/mappers/explore_state_mapper.dart
@@ -5,10 +5,9 @@ import 'package:seeds/v2/datasource/remote/model/voice_model_campaign.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/domain-shared/result_to_state_mapper.dart';
 import 'package:seeds/v2/screens/explore/interactor/viewmodels/explore_state.dart';
-import 'package:seeds/v2/screens/profile/interactor/usecases/get_profile_use_case.dart';
+import 'package:seeds/v2/screens/profile_screens/profile/interactor/usecases/get_profile_use_case.dart';
 
 class ExploreStateMapper extends StateMapper {
-
   ExploreState mapResultsToState(ExploreState currentState, List<Result> results) {
     if (areAllResultsError(results)) {
       return currentState.copyWith(pageState: PageState.failure, errorMessage: 'Error Loading Page');
