+++ lib/v2/screens/profile_screens/settings/interactor/usecases/get_profile_use_case.dart
@@ -1,0 +1,12 @@
+import 'package:async/async.dart';
+import 'package:seeds/v2/datasource/remote/api/profile_repository.dart';
+
+export 'package:async/src/result/result.dart';
+
+class GetProfileUseCase {
+  final ProfileRepository _profileRepository = ProfileRepository();
+
+  Future<Result> run() {
+    return _profileRepository.getProfile();
+  }
+}
