+++ lib/v2/screens/profile_screens/settings/set_currency/interactor/viewmodels/set_currency_state.dart
